Moved pages into a new "/pages" folder
